+++ src/TrashLib.TestLibrary/NewCf.cs
@@ -9,7 +9,7 @@ public static class NewCf
     public static CustomFormatData Data(string name, string trashId, int? score = null)
     {
         var json = JObject.Parse($"{{'name':'{name}'}}");
-        return new CustomFormatData(name, trashId, score, new JObject(json));
+        return new CustomFormatData("", name, trashId, score, new JObject(json));
     }
 
     public static ProcessedCustomFormatData Processed(string name, string trashId, int? score = null)
@@ -19,7 +19,7 @@ public static class NewCf
 
     public static ProcessedCustomFormatData Processed(string name, string trashId, int? score, JObject json)
     {
-        return new ProcessedCustomFormatData(new CustomFormatData(name, trashId, score, json));
+        return new ProcessedCustomFormatData(new CustomFormatData("", name, trashId, score, json));
     }
 
     public static ProcessedCustomFormatData Processed(string name, string trashId, JObject json)
@@ -38,7 +38,7 @@ public static class NewCf
     public static ProcessedCustomFormatData Processed(string name, string trashId, JObject json,
         TrashIdMapping? cacheEntry)
     {
-        return new ProcessedCustomFormatData(new CustomFormatData(name, trashId, null, json))
+        return new ProcessedCustomFormatData(new CustomFormatData("", name, trashId, null, json))
         {
             CacheEntry = cacheEntry
         };