+++ src/TrashLib.Tests/CustomFormat/Guide/CustomFormatLoaderTest.cs
@@ -1,3 +1,4 @@
+using System.IO.Abstractions;
 using System.IO.Abstractions.Extensions;
 using System.IO.Abstractions.TestingHelpers;
 using FluentAssertions;
@@ -20,19 +21,22 @@ public class CustomFormatLoaderTest : IntegrationFixture
         var sut = Resolve<ICustomFormatLoader>();
         Fs.AddFile("first.json", new MockFileData("{'name':'first','trash_id':'1'}"));
         Fs.AddFile("second.json", new MockFileData("{'name':'second','trash_id':'2'}"));
+        Fs.AddFile("collection_of_cfs.md", new MockFileData(""));
 
-        var results = sut.LoadAllCustomFormatsAtPaths(new[] {Fs.CurrentDirectory()});
+        var dir = Fs.CurrentDirectory();
+        var results = sut.LoadAllCustomFormatsAtPaths(new[] {dir}, dir.File("collection_of_cfs.md"));
 
         results.Should().BeEquivalentTo(new[]
         {
-            NewCf.Data("first", "1"),
-            NewCf.Data("second", "2")
+            NewCf.Data("first", "1") with {FileName = "first.json"},
+            NewCf.Data("second", "2") with {FileName = "second.json"}
         });
     }
 
     [Test]
     public void Trash_properties_are_removed()
     {
+        Fs.AddFile("collection_of_cfs.md", new MockFileData(""));
         Fs.AddFile("first.json", new MockFileData(@"
 {
   'name':'first',
@@ -44,12 +48,14 @@ public class CustomFormatLoaderTest : IntegrationFixture
 
         var sut = Resolve<ICustomFormatLoader>();
 
-        var results = sut.LoadAllCustomFormatsAtPaths(new[] {Fs.CurrentDirectory()});
+        var dir = Fs.CurrentDirectory();
+        var results = sut.LoadAllCustomFormatsAtPaths(
+            new[] {dir}, dir.File("collection_of_cfs.md"));
 
         const string expectedExtraJson = @"{'name':'first','extra': 'e1'}";
 
         results.Should()
-            .ContainSingle().Which.ExtraJson.Should()
+            .ContainSingle().Which.Json.Should()
             .BeEquivalentTo(JObject.Parse(expectedExtraJson), op => op.Using(new JsonEquivalencyStep()));
     }
 }