+++ src/TrashLib/Services/CustomFormat/Guide/CustomFormatCategoryItem.cs
@@ -1,0 +1,3 @@
+namespace TrashLib.Services.CustomFormat.Guide;
+
+public record CustomFormatCategoryItem(string CategoryName, string CfName, string CfAnchor);