+++ src/TrashLib/Services/CustomFormat/Guide/ICustomFormatCategoryParser.cs
@@ -1,0 +1,8 @@
+using System.IO.Abstractions;
+
+namespace TrashLib.Services.CustomFormat.Guide;
+
+public interface ICustomFormatCategoryParser
+{
+    ICollection<CustomFormatCategoryItem> Parse(IFileInfo collectionOfCustomFormatsMdFile);
+}