+++ src/TrashLib/Services/CustomFormat/Guide/ICustomFormatLoader.cs
@@ -5,5 +5,6 @@ namespace TrashLib.Services.CustomFormat.Guide;
 
 public interface ICustomFormatLoader
 {
-    ICollection<CustomFormatData> LoadAllCustomFormatsAtPaths(IEnumerable<IDirectoryInfo> jsonPaths);
+    ICollection<CustomFormatData> LoadAllCustomFormatsAtPaths(IEnumerable<IDirectoryInfo> jsonPaths,
+        IFileInfo collectionOfCustomFormats);
 }