+++ src/TrashLib/Services/CustomFormat/Models/CustomFormatData.cs
@@ -4,8 +4,12 @@ using Newtonsoft.Json.Linq;
 namespace TrashLib.Services.CustomFormat.Models;
 
 public record CustomFormatData(
+    string FileName,
     string Name,
     string TrashId,
     int? Score,
-    [property: JsonExtensionData] JObject ExtraJson
-);
+    [property: JsonExtensionData] JObject Json
+)
+{
+    public string? Category { get; init; }
+}